CRUD api backend completed

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 import expressAsyncHandler from 'express-async-handler'
 import { Goal } from '../models/goalModel.js'
+import { User } from '../models/userModel.js'
 
 // @desc Get goals
 // @route GET /api/goals
 // @access Private
 const getGoals = expressAsyncHandler(async (req, res) => {
-  const goals = await Goal.find()
+  const goals = await Goal.find({ user: req.user.id })
   res.status(200).json(goals)
 })
 
@@ -17,8 +18,8 @@ const setGoal = expressAsyncHandler(async (req, res) => {
     res.status(400)
     throw new Error('Please add a text field')
   }
-  const goal = await Goal.create({ text: req.body.text })
-  res.status(200).json(goal)
+  const goal = await Goal.create({ text: req.body.text, user: req.user.id })
+  res.status(201).json(goal)
 })
 
 // @desc Update goals
@@ -29,6 +30,19 @@ const updateGoal = expressAsyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400)
     throw new Error('Goal not found')
+  }
+
+  const user = await User.findById(req.user.id)
+  // check for user
+  if (!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  // make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorised')
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -46,6 +60,20 @@ const deleteGoal = expressAsyncHandler(async (req, res) => {
     res.status(400)
     throw new Error('Goal not found')
   }
+
+  const user = await User.findById(req.user.id)
+  // check for user
+  if (!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  // make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorised')
+  }
+
   await goal.deleteOne()
   res.status(200).json({ id: req.params.id })
 })
